corrected page routing
